test max add ok

+++ FOAD/JS/Exo/Point/Point.js
@@ -29,8 +29,8 @@ class Point {
      */
 
     constructor(_x, _y) {
-        this.x = parseFloat(_x || 0);
-        this.y = parseFloat(_y || 0);
+        this.x = parseInt(_x || 0);
+        this.y = parseInt(_y || 0);
     }
 
     /**
